refactor(stella-pipeline): move the working-tree observation cluster into verify_probes
Pure relocation, paving #1553's fix: gather_diff, absorb_probe,
observed_mutations, untracked_added_lines, verification_honest_diff,
the DIFF_PROBE markers, DiffProbe and its concurrency bound move from
pipeline.rs into pipeline/verify_probes.rs — the module whose charter
is already post-execution observation. Call sites are unchanged (the
methods stay on Pipeline, pub(super) covers the module tree); the one
test import updates its path.

Refs #1553
