+++ crates/stella-pipeline/src/pipeline/verify_probes.rs
@@ -1,8 +1,12 @@
 //! Post-execution observation probes for [`super::Pipeline`] — the
 //! touched-tests observation that feeds the flip oracle (#860 typed
-//! outcomes) and the lint delta behind the regression veto (#861).
+//! outcomes), the lint delta behind the regression veto (#861), and the
+//! working-tree diff observation (`gather_diff`/`absorb_probe`) with the
+//! honest-empty-diff discipline around it.
 //! A child module of `pipeline` so the private candidate types stay
 //! reachable, split out to keep the orchestrator under the size gate.
+
+use futures_util::StreamExt as _;
 
 use super::*;
 
@@ -207,4 +211,269 @@ impl<'a> Pipeline<'a> {
             diff_coverage: Some(inputs.diff_coverage.as_str().to_string()),
         }
     }
+
+    /// The real added-line count of an untracked file, via a no-index diff
+    /// numstat (`<added>\t<deleted>\t<path>`). A binary file numstats as `-`
+    /// and counts as one changed line (a change the ladder must see, but
+    /// unmeasurable in lines). Counting real lines — not a flat 1 per file —
+    /// is what keeps a large untracked file from slipping under the diff budget
+    /// and taking `SubmitFast`. A single file's numstat is one line, so this is
+    /// safe against the diagnostic runner's output truncation.
+    async fn untracked_added_lines(&self, surface: CandidateSurface<'_>, path: &str) -> u32 {
+        let out = surface
+            .diagnostics
+            .run_diagnostic(&DiagnosticInvocation::UntrackedNumstat {
+                path: path.to_string(),
+            })
+            .await;
+        out.stdout_tail
+            .lines()
+            .find(|l| !l.trim().is_empty())
+            .and_then(|l| l.split('\t').next())
+            .and_then(|added| added.trim().parse::<u32>().ok())
+            .unwrap_or(1)
+    }
+
+    /// Run the diff command and return `(changed_line_count, raw_diff)`.
+    ///
+    /// `git diff` cannot see untracked files, so a turn whose entire change is
+    /// a NEW (or edited-untracked) file would read as "no diff" — skipping
+    /// verification via the zero-diff guard and showing the verifier an empty
+    /// diff. When the configured command is git's, untracked files that were
+    /// **created or modified this turn** — present now with either no
+    /// `untracked_before` entry or a changed fingerprint — are appended with
+    /// their real added-line counts. Untouched dirty files (same fingerprint)
+    /// are excluded, so pre-existing state is never attributed to the turn, and
+    /// a large untracked file cannot slip under the diff-size budget.
+    pub(super) async fn gather_diff(
+        &self,
+        surface: CandidateSurface<'_>,
+        untracked_before: &HashMap<String, String>,
+    ) -> DiffProbe {
+        let Some(diagnostic) = &self.config.diff_diagnostic else {
+            // No probe configured is not "nothing changed" either: this host
+            // simply has no diff channel, so it reports one fewer way to see.
+            return DiffProbe {
+                lines: 0,
+                text: String::new(),
+                available: false,
+            };
+        };
+        let out = surface.diagnostics.run_diagnostic(diagnostic).await;
+        // `git diff` exits 0 on a clean tree, so a non-zero exit with nothing on
+        // stdout is the machinery failing to READ the tree, never a report that
+        // the tree is unchanged. Observed in the wild as a candidate shadow
+        // whose worktree registration was gone by the time it was probed
+        // (`git add -A` → "fatal: not a git repository"). See
+        // [`DIFF_PROBE_FAILED`] for why the empty string could not be left to
+        // `verification_honest_diff` to catch. Scoped to `GitDiff` because
+        // `--no-index --numstat` exits 1 whenever the files differ, which is
+        // that probe's ordinary success.
+        if matches!(diagnostic, DiagnosticInvocation::GitDiff)
+            && !out.passed()
+            && out.stdout_tail.trim().is_empty()
+        {
+            // git's own words, matched case-insensitively because the phrasing
+            // is stable across versions but its capitalization is not.
+            let not_a_repo = out
+                .stderr_tail
+                .to_ascii_lowercase()
+                .contains("not a git repository");
+            return DiffProbe {
+                lines: 0,
+                text: if not_a_repo {
+                    DIFF_PROBE_NOT_A_REPO.to_string()
+                } else {
+                    DIFF_PROBE_FAILED.to_string()
+                },
+                available: false,
+            };
+        }
+        let mut lines = count_diff_lines(&out.stdout_tail);
+        let mut text = out.stdout_tail;
+        if matches!(diagnostic, DiagnosticInvocation::GitDiff) {
+            let after = surface.repo_status.untracked_fingerprints().await;
+            // Created (absent before) OR modified (fingerprint changed) this
+            // turn — never an untouched dirty file.
+            let mut fresh: Vec<&str> = after
+                .iter()
+                .filter(|(path, fp)| untracked_before.get(*path) != Some(*fp))
+                .map(|(path, _)| path.as_str())
+                .collect();
+            fresh.sort(); // deterministic order for the appended evidence
+            // Each of these is a `git diff --no-index --numstat` subprocess.
+            // Run sequentially, a turn that creates many untracked files paid
+            // one full process round-trip per file — on every verification
+            // observation and again after every revision, so the cost is
+            // repaid per revision per candidate.
+            //
+            // Bounded concurrency rather than a truncating cap: this count
+            // feeds the zero-diff guard and the diff-size budget, so dropping
+            // the tail would let a large untracked change slip under a budget
+            // it should have tripped. `buffered` preserves input order, so
+            // the appended evidence stays deterministic.
+            let counted: Vec<(&str, u32)> =
+                futures_util::stream::iter(fresh.into_iter().map(|path| async move {
+                    (path, self.untracked_added_lines(surface, path).await)
+                }))
+                .buffered(UNTRACKED_NUMSTAT_CONCURRENCY)
+                .collect()
+                .await;
+            for (path, added) in counted {
+                lines += added;
+                // The shared builder, so `witness::warrant` can parse these
+                // paths back out and hold them to the same path rules as
+                // tracked changes — a hand-rolled format here silently
+                // blinded the warrant to every untracked file.
+                text.push('\n');
+                text.push_str(&crate::witness::warrant::untracked_change_line(path, added));
+            }
+        }
+        DiffProbe {
+            lines,
+            text,
+            available: true,
+        }
+    }
+
+    /// Mutating file touches this candidate has made, read from the recorder
+    /// that emitted the `FileChange` events rather than from a wrapper around
+    /// the engine's sender.
+    ///
+    /// This is the fix for the counter that read `0` while six `file_change`
+    /// events sat in the very stream the verifier's run produced (#973). The two
+    /// numbers came from different wires: `ToolRegistry::record_touch` sends to
+    /// the channel the *host* attached, and the tally lived on a sender the
+    /// pipeline handed the *engine*, which no file tool ever uses.
+    ///
+    /// `max` of the two, because they are independent lower bounds rather than
+    /// a sum: the recorder's delta is authoritative wherever a `FileTouchPort`
+    /// is wired, and the event tally is the only signal for a host with no
+    /// recorder at all. Neither can double-count the other — the sender the
+    /// tally wraps is built per-turn inside this pipeline and never handed out.
+    fn observed_mutations(&self, state: &CandidateState) -> u32 {
+        let delta = self
+            .touches
+            .mutations_recorded()
+            .saturating_sub(state.touch_baseline);
+        u32::try_from(delta)
+            .unwrap_or(u32::MAX)
+            .max(state.file_changes)
+    }
+
+    /// Fold one working-tree observation into `state`, refreshing every
+    /// evidence channel from its own source at the same instant.
+    ///
+    /// One function, so the channels cannot be updated apart and disagree about
+    /// which round they describe — the honest-diff text in particular is built
+    /// from the touch count, and reading a stale one is how a real change gets
+    /// rendered as an empty diff.
+    pub(super) fn absorb_probe(&self, state: &mut CandidateState, probe: DiffProbe) {
+        // Attribute shell-driven changes BEFORE the count is read. The
+        // recorder only sees what a tool's input declared, and `bash` declares
+        // nothing, so without this the count below is blind to the tool that
+        // does most of the work on Terminal-Bench. Settling first is what
+        // makes `file_changes` an honest answer rather than a CRUD-tool tally.
+        // Settling also re-arms, from the same walk, so a revision's changes
+        // are bracketed exactly the way this one's were without paying for a
+        // second walk of an identical tree.
+        self.touches.settle_workspace_probe();
+        state.file_changes = self.observed_mutations(state);
+        // The authored channel is read AFTER settling, for the same reason the
+        // count is: the probe's attributions are recorded through the same
+        // emitter, so reading first would describe the turn before its opaque
+        // calls had been accounted for.
+        let authored = self.touches.authored_diff();
+        // `max`, not a sum — these are two independent LOWER BOUNDS on one
+        // number, not two disjoint tallies. A tracked file edited through
+        // `write_file` is counted by both, and adding them would double it.
+        // (The same reasoning as `observed_mutations` above.)
+        state.diff_lines = probe.lines.max(authored.lines);
+        // Deliberately NOT widened by the authored channel. `diff_available`
+        // means "the configured probe could read the working tree", and the
+        // ladder's blind-evidence rung is built on that exact claim. An
+        // authored diff proves what was written, never what survived to sit on
+        // disk at the end of the turn — so it informs the verifier without
+        // asserting that the tree was successfully re-read.
+        state.diff_available = probe.available;
+        state.diff_text = verification_honest_diff(
+            authored::splice_authored(probe.text, &authored),
+            state.file_changes,
+        );
+    }
 }
+
+/// An empty diff is ambiguous: it can mean the agent genuinely changed
+/// nothing, OR that the diff machinery is blind to changes that DID happen
+/// (work that was committed, a baseline mismatch, an uncaptured file). Handed
+/// a bare empty string, a verifier reads the second case as the first and
+/// concludes "no changes were made" — the archetypal verification lie that
+/// once drove an agent to reinitialize git to beat the check.
+///
+/// The pipeline already has the independent signal that resolves the
+/// ambiguity: `file_changes`, the count of `FileChange` events the turn
+/// emitted. When that is positive but the diff is empty, the honest report is
+/// "the tree changed but the diff could not be captured" — a *couldn't
+/// verify*, never a *verified nothing*. Every consumer (verifier, guidance,
+/// evidence) then sees the truth.
+pub(super) fn verification_honest_diff(diff_text: String, file_changes: u32) -> String {
+    if file_changes > 0 && diff_text.trim().is_empty() {
+        format!(
+            "[{file_changes} file-change event(s) were observed this turn, but the diff could \
+             not be captured — the change is real; the diff is blind to it. This is NOT evidence \
+             that nothing changed; verify the result on its own merits.]"
+        )
+    } else {
+        diff_text
+    }
+}
+
+/// What [`Pipeline::gather_diff`] reports when the diff probe *failed* rather
+/// than found nothing.
+///
+/// `verification_honest_diff` resolves the same ambiguity from `file_changes`,
+/// but that signal is structurally unavailable to an isolated candidate: the
+/// engine emits no `FileChange` events (the deck's `FileChangeTap` synthesizes
+/// them, and it wraps only the SESSION tool stack), so inside a best-of-N or
+/// witness candidate the count is always zero. A candidate whose probe went
+/// blind therefore handed an empty string to [`crate::witness::warrant`], which
+/// read it as [`crate::NoWitnessReason::NothingChanged`] and completed the run
+/// with a PASSING verdict stating no behavior had changed — the exact
+/// verification lie the honest-diff guard exists to prevent, reached with no
+/// witness stage and no warning. Non-empty text keeps the warrant fail-closed
+/// on the one path its own guard could not see.
+const DIFF_PROBE_FAILED: &str = "[the diff probe failed; the working tree could not be read. This \
+     is NOT evidence that nothing changed — verify the result on its own merits.]";
+
+/// The same report, for the one cause worth naming separately: there is no git
+/// repository here, so `git diff` can never answer.
+///
+/// Terminal-Bench task images are plain directories. `DIFF_PROBE_FAILED` reads
+/// as a transient fault a reader might expect to clear on a retry; this one
+/// says the probe is structurally inapplicable to this workspace, which is the
+/// difference between "try again" and "use another channel" (#973).
+const DIFF_PROBE_NOT_A_REPO: &str = "[the working tree is not a git repository, so the diff probe cannot read it at all. This is \
+     a permanent property of this workspace, NOT evidence that nothing changed — verify the \
+     result on its own merits.]";
+
+/// One observation of the working tree by [`Pipeline::gather_diff`].
+///
+/// `available` is the field that did not exist and had to: without it, "the
+/// probe read the tree and it was unchanged" and "the probe could not read the
+/// tree" both arrived as `(0, ...)`, and every consumer downstream was free to
+/// read the second as the first. The text has said the right thing for a while;
+/// nothing could act on it, because prose is not a signal a ladder can branch
+/// on.
+pub(super) struct DiffProbe {
+    pub(super) lines: u32,
+    pub(super) text: String,
+    /// Whether the probe could read the working tree AT ALL. Never `false`
+    /// merely because the diff came back empty.
+    pub(super) available: bool,
+}
+
+/// How many `git diff --no-index --numstat` probes [`Pipeline::gather_diff`]
+/// keeps in flight at once. High enough that the usual handful of new files
+/// costs roughly one round-trip instead of N, low enough that a turn which
+/// creates hundreds cannot fork an unbounded burst of git processes.
+const UNTRACKED_NUMSTAT_CONCURRENCY: usize = 16;
